logwowanie metryk pod strojenie

+++ LEM/src/dv_control/include/Vec2.hpp
@@ -4,11 +4,11 @@
 namespace v2_control
 {
 struct Vec2 {
-    float x;
-    float y;
+    double x;
+    double y;
 
     Vec2() : x(0.0f), y(0.0f) {}
-    Vec2(float xx, float yy) : x(xx), y(yy) {}
+    Vec2(double xx, double yy) : x(xx), y(yy) {}
 
     inline Vec2 operator+(const Vec2& other) const {
         return Vec2(x + other.x, y + other.y);
@@ -18,17 +18,17 @@ struct Vec2 {
         return Vec2(x - other.x, y - other.y);
     }
 
-    inline Vec2 operator*(float scalar) const {
+    inline Vec2 operator*(double scalar) const {
         return Vec2(x * scalar, y * scalar);
     }
 
-    inline Vec2& operator/=(float scalar) {
+    inline Vec2& operator/=(double scalar) {
         x /= scalar;
         y /= scalar;
         return *this;
     }
 
-    inline float length() const {
+    inline double length() const {
         return std::sqrt(x * x + y * y);
     }
 };
